+++ static_client/react-app/src/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './css/index.css';
 import App from './App';
-import MusicalAccompaniment from './MusicalAccompaniment';
-import LineHat from './LineHat';
-import GetLine from './GetLine';
+import MusicalAccompaniment from './music/MusicalAccompaniment';
+import LineHat from './line_hat/LineHat';
+import GetLine from './line_hat/GetLine';
 import registerServiceWorker from './registerServiceWorker';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'; 
-import AudiencePlaySong from './AudiencePlaySong';
-import AdminMusic from './AdminMusic';
-import SoloNewChoice from './SoloNewChoice';
-import AudienceNewChoice from './AudienceNewChoice';
-import AdminNewChoice from './AdminNewChoice';
-import HatAdmin from './HatAdmin';
+import AudiencePlaySong from './music/AudiencePlaySong';
+import AdminMusic from './music/AdminMusic';
+import SoloNewChoice from './new_choice/SoloNewChoice';
+import AudienceNewChoice from './new_choice/AudienceNewChoice';
+import AdminNewChoice from './new_choice/AdminNewChoice';
+import HatAdmin from './line_hat/HatAdmin';
 
 
 ReactDOM.render(
