+++ static_client/react-app/src/line_hat/LineHat.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './linehat.css';
+import '../css/linehat.css';
 import axios from 'axios';
