+++ static_client/react-app/src/new_choice/SoloNewChoice.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import new_choice from './songs/newChoice.mp3';
+import new_choice from '../music/songs/newChoice.mp3';
 
 class SoloNewChoice extends Component {
 	constructor() {
